requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.11.0
fpdf2>=2.7.0
numpy>=1.21.0